module7-task2 - переместил работу со списками и городами внутрь компонента MainPage
